Moving fetch logic to child component and added the search functionality
